WIP: Working responsive nav styles
