+++ src/main.scss
@@ -157,7 +157,6 @@ $enable-dark-mode: true;
 @import "./components/bs5/inpagenav/inpagenav.scss";
 @import "./components/bs5/sidenav/sidenav.scss"; //includes side-nav.scss
 @import "./components/bs5/navbar/navbar.scss";
-@import "./components/bs5/navbarBS5/navbar.bs5.scss";
 @import "./components/bs5/table/table.scss";
 @import "./components/bs5/pagination/pagination.scss";
 @import "./components/bs5/typography/typography.scss";
